Make the sessions service more rigorous

+++ app/tabular/services/sessions.rb
@@ -5,7 +5,9 @@ module Tabular
       module_function
 
       # If the given credentials are valid, create a session.
-      def login!(user, password)
+      def login!(username, password)
+        user = Models::User.find_by(username: username)
+        fail Errors::NoSuchModel unless user
         id, salt, hash = user.id, user.password_salt, user.password_hash
         Passwords.authenticate!(password, salt, hash)
         Models::Session.create!(key: unique_session_keys.next, user_id: id)
@@ -13,6 +15,7 @@ module Tabular
 
       # Destroy all the sessions with the given key.
       def logout!(key)
+        fail Errors::Unauthorized unless Models::Session.exists?(key: key)
         Models::Session.destroy_all(key: key)
       end
